k8s/arangodb/jwt_bearer_token/requirements.txt
kubernetes >= 11.0.0
pyjwt >= 1.7.1
